Migrate 'muxer' HTTP router to external repository
This commit marks the transition of 'muxer', my lightweight HTTP router
written in Go, from an internal development repository to a publicly
accessible external repository. 'Muxer' has been a core component in my
production environments for over a year, demonstrating its reliability and
efficiency in handling HTTP routing.

The design of 'muxer' is centered around simplicity and flexibility, with
an API that's intuitive for users familiar with the net/http package in Go.
It supports straightforward registration of HTTP handlers and middleware,
and efficiently matches incoming requests to their respective handlers
based on path and HTTP method.

Moving to an external repository is a strategic step to encourage open
source collaboration. By opening 'muxer' to the public, I aim to foster
community contributions, allowing it to benefit from the collective
experience and expertise of the broader Go developer community. This shift
also aligns with my commitment to open source principles and transparent
development.

The entire history of 'muxer', including all features and bug fixes
developed over the past year, is included in this migration. This ensures
that new contributors can understand the evolution of the project and
contribute more effectively.

I anticipate that this move will not only enhance the visibility of 'muxer'
but also accelerate its development, benefiting both current and future
users of the router.

By making 'muxer' open source, I embrace the potential for more innovative
uses and improvements, ensuring its continued relevance and effectiveness
as a lightweight HTTP routing solution.

While I eagerly await the changes in the stdlib that add some of this
functionality, I have yet to evaluate everything it is providing and if
there is enough feature parity I will probably archive this in favor of
stdlib as who wants dependencies when they don't have to.

Until then, happy coding friends!!